Fixed issues in how i resolve resources, fixed issue in deviceio, and in devicemanager

+++ modules/arch/x86/ps2/port.c
@@ -317,7 +317,7 @@ PS2PortInitialize(
     }
 
     // Initialize interrupt resources
-    RegisterFastInterruptMemoryResource(&Port->Interrupt, Port, sizeof(PS2Port_t), 0);
+    RegisterFastInterruptMemoryResource(&Port->Interrupt, (uintptr_t)Port, sizeof(PS2Port_t), 0);
     RegisterInterruptContext(&Port->Interrupt, Port);
 
     // Start out by doing an interface
